Create searching_an_element_in_a_matrix.cpp
This is one of the most asked questions in an interview.
problem description-Given an n x n matrix and an integer x, find the position of x in the matrix if it is present. Otherwise, print “Element not found”. Every row and column of the matrix is sorted in increasing order. The designed algorithm should have linear time complexity